refactor: split cache models.py into ONTAP REST API category modules
